CHANGE: MSVC 2015 compatibility (Dr. Supergood, from Standard) CHANGE:	Use timeGetTime for windows Posix builds too (Prissi, from Standard) CHANGE: Use ShGetFolderPAth for user directory (Prisi, from Standard) CHANGE: Add Hiaku to find user directory (Prissi, from Standard)

+++ simsys.h
@@ -117,7 +117,7 @@ void ex_ord_update_mx_my();
 void GetEvents();
 void GetEventsNoWait();
 
-unsigned long dr_time();
+uint32 dr_time();
 void dr_sleep(uint32 millisec);
 
 // error message in case of fatal events
